# routes/admin/v1/admin.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FG Digital",
        "description": "ADMIN APIs - Project Management Portal",
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/admin/v1",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
...
    "paths": {
...
        "/employee/add": {
            "post": {
                "tags": ["5. Employee"],
                "summary": "Add New Employee",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["email", "password", "name"],
                        "properties": {
                            "email": { "type": "string" },
                            "password": { "type": "string" },
                            "name": { "type": "string" }
                        },
                        "example": {
                            "email": "test@gmail.com",
                            "password": "my-secret-password",
                            "name": "Employee PQR"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/employee/get": {
            "get": {
                "tags": ["5. Employee"],
                "summary": "Get Employee Details",
                "parameters": [{
                    "in": "query",
                    "name": "employee_id",
                    "description": "Employee ID",
                    "required": false,
                    "items": { "type": "string" }
                }],
                "responses": {}
            }
        },
        "/employee/update/{employee_id}": {
            "post": {
                "tags": ["5. Employee"],
                "summary": "Update existing employee",
                "consumes": ["multipart/form-data"],
                "parameters": [{
                    "in": "path",
                    "name": "employee_id",
                    "description": "Employee ID",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "example": 20201
                    }
                }, {
                    "in": "formData",
                    "name": "name",
                    "required": false,
                    "type": "string",
                    "description": "Employee Name",
                    "schema": {
                        "example": "John Patel"
                    }
                }, {
                    "in": "formData",
                    "name": "position",
                    "description": "Employee Position (Ex. Website Developer, Marketing Specialist, Application Developer, Content Writer, SEO Specialist, Social Media Manager, Graphics Designer, Videographer, Video Editor )",
                    "required": false,
                    "type": "string",
                    "schema": {
                        "example": "Website Developer"
                    }
                }, {
                    "in": "formData",
                    "name": "skills",
                    "description": "Employee Skills",
                    "required": false,
                    "type": "string",
                    "schema": {
                        "example": "HTML, Video Editing, PHP, JavaScript"
                    }
                }, {
                    "in": "formData",
                    "name": "profile_image",
                    "description": "Employee Profile Image",
                    "required": false,
                    "type": "file"
                }, {
                    "in": "formData",
                    "name": "email",
                    "description": "Employee Email Address",
                    "required": false,
                    "type": "string",
                    "schema": {
                        "example": "John123@gmail.com"
                    }
                }],
                "responses": {}
            }
        },
        "/employee/change-password": {
            "post": {
                "tags": ["5. Employee"],
                "summary": "Change Employee Password",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["employee_id"],
                        "properties": {
                            "employee_id": { "type": "string" },
                            "password": { "type": "string" }
                        },
                        "example": {
                            "employee_id": "20201",
                            "password": "my-secret-password"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/employee/remove": {
            "get": {
                "tags": ["5. Employee"],
                "summary": "Remove existing employee",
                "parameters": [{
                    "in": "query",
                    "name": "employee_id",
                    "type": "integer",
                    "description": "Employee ID",
                    "required": true,
                    "schema": {
                        "example": 20201
                    }
                }],
                "responses": {}
            }
        },
        "/employee/social-account/add": {
            "post": {
                "tags": ["5. Employee"],
                "summary": "Add social media account",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["employee_id", "account_name", "reference"],
                        "properties": {
                            "employee_id": { "type": "integer", "description": "Get ID using /employee/get" },
                            "account_name": { "type": "string" },
                            "reference": { "type": "string" }
                        },
                        "example": {
                            "employee_id": 20201,
                            "account_name": "Facebook",
                            "reference": "ABC"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/employee/social-account/update": {
            "post": {
                "tags": ["5. Employee"],
                "summary": "Update existing social account details",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["employee_id", "account_id"],
                        "properties": {
                            "employee_id": { "type": "integer", "description": "Get ID using /employee/get" },
                            "account_id": { "type": "string", "description": "Get Account ID using /employee/get" },
                            "account_name": { "type": "string" },
                            "reference": { "type": "string" }
                        },
                        "example": {
                            "employee_id": 20201,
                            "account_id": "-MDEEyyvvIQLvPhmLzZa",
                            "account_name": "Facebook",
                            "reference": "ABC"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/employee/social-account/remove": {
            "get": {
                "tags": ["5. Employee"],
                "summary": "Remove existing social account",
                "parameters": [{
                    "in": "query",
                    "name": "employee_id",
                    "type": "integer",
                    "description": "Employee ID",
                    "required": true,
                    "schema": {
                        "example": 20201
                    }
                }, {
                    "in": "query",
                    "name": "account_id",
                    "type": "string",
                    "description": "Social Account ID",
                    "required": true,
                    "schema": { "example": "-MDUWHxmzI4N9LFBDpU_" }
                }],
                "responses": {}
            }
        },
...
    }
}